Reorganize files: vector tile pre-processing in sources/ subdirectory
